Enhance spell mechanics and refactor damage handling
Refactored `RecieveSpellAttack` in `EntityAttack.cs` to support binary spells, hit chance calculations, and improved damage processing. Introduced `SpellEffect` as a parameter for better context.

Added logic for binary and non-binary spell resistance, including resist messages and damage type handling. Updated `ProcessDamage` to use `SpellEffect.Name` and added TODOs for future improvements.

Modified constructors in `OverTime` and `Instant` to include `isBinary` and `spellSchools` parameters. Refactored `SpellEffect` to support binary handling and spell schools.

Replaced `double` with `float` in `ClassData`, `Attack`, and other stats-related classes for better performance. Updated `Spell` to use `MathF` for floating-point calculations.

Renamed `SpellResitance` to `SpellResistance` across the codebase for consistency. Improved resistance calculations for binary and non-binary spells using `HashSet<SpellSchool>`.

Refactored `ApplyCrushingDamage` to apply to all damage types. Added TODOs for handling dual-wield penalties for spells and shield-specific damage blocking.

These changes improve spell mechanics, enhance code clarity, and prepare the codebase for future extensions.

+++ Project 1/GameObjects/Unit/Classes/ClassData.cs
@@ -105,8 +105,8 @@ namespace Project_1.GameObjects.Unit.Classes
         //Rogues receive 1% Critical Strike Chance for every 29 points of Agility.
         //Hunters receive 1% Critical Strike Chance for every 53 points of Agility.
 
-        public double AttackCritChanceScaler => attackCritChanceScaler;
-        double attackCritChanceScaler;
+        public float AttackCritChanceScaler => attackCritChanceScaler;
+        float attackCritChanceScaler;
 
 
         //        You gain Critical Strike chance at varying points, depending on your class:
@@ -118,18 +118,18 @@ namespace Project_1.GameObjects.Unit.Classes
         //Paladins receive 1% Spell Critical Strike chance for every 54 points of intellect.
 
 
-        public double SpellCritChanceScaler => spellCritChanceScaler;
-        double spellCritChanceScaler;
+        public float SpellCritChanceScaler => spellCritChanceScaler;
+        float spellCritChanceScaler;
 
         //All classes but Hunters and Rogues receive 1% Dodge for every 20 points of Agility.
         //Rogues receive 1% Dodge for every 14.5 points of Agility.
         //Hunters receive 1% Dodge for every 26 points of Agility.
 
-        public double DodgeChanceScaler => dodgeChanceScaler;
-        double dodgeChanceScaler;
+        public float DodgeChanceScaler => dodgeChanceScaler;
+        float dodgeChanceScaler;
 
-        public double BaseDodge => baseDodge;
-        double baseDodge;
+        public float BaseDodge => baseDodge;
+        float baseDodge;
 
         //Class	    Base dodge	AGI:Dodge ratio
         //Druid	    0.9%	    20
